requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.10
django-filter==23.5
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
Markdown==3.5.2
packaging==23.2
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
zipp==3.17.0
